Some modifications
Ex: Remove ExpansionList
Change name BootcampMagicTests and BootcampMagicUITests
