+++ Bootcamp-Magic/Model/Expansion.swift
@@ -11,3 +11,19 @@ struct Expansion: Decodable {
     let code: String
     let name: String
 }
+
+struct ExpansionList: Decodable {
+    
+    let expansions: [Expansion]
+    
+    init(from decoder: Decoder) throws {
+        let container = try decoder.container(keyedBy: CodingKeys.self)
+        expansions = try container.decode([Expansion].self, forKey: .sets)
+    }
+}
+
+extension ExpansionList {
+    enum CodingKeys: String, CodingKey {
+        case sets
+    }
+}
