+++ Bootcamp-MagicTests/Bootcamp_MagicTests.swift
@@ -8,7 +8,7 @@
 import XCTest
 @testable import Bootcamp_Magic
 
-class Bootcamp_MagicTests: XCTestCase {
+class BootcampMagicTests: XCTestCase {
 
     override func setUpWithError() throws {
         // Put setup code here. This method is called before the invocation of each test method in the class.
