+++ Bootcamp-MagicUITests/Bootcamp_MagicUITests.swift
@@ -7,7 +7,7 @@
 
 import XCTest
 
-class Bootcamp_MagicUITests: XCTestCase {
+class BootcampMagicUITests: XCTestCase {
 
     override func setUpWithError() throws {
         // Put setup code here. This method is called before the invocation of each test method in the class.
